Restructures the main section and replaces divs with semantic tags...
I am having a hard time reading the add -p output for this commit
so I just added it all at once. The main content is split into five
rows. The first is the date and theme. Next is the Stay in Touch
header, which was added to break apart the contact and meetup
details. Then there is a row with two columns for the meetup link
and the matrix details. Finally is a row for the SIB form.